+++ Demo/Demo/Models/ShippingPreference.swift
@@ -1,7 +1,0 @@
-enum ShippingPreference: String, CaseIterable, Identifiable {
-    case noShipping = "NO_SHIPPING"
-    case providedAddress = "SET_PROVIDED_ADDRESS"
-    case getFromFile = "GET_FROM_FILE"
-
-    var id: ShippingPreference { self }
-}