+++ UnitTests/PayPalWebPaymentsTests/PayPalWebCheckoutClient_Tests.swift
@@ -26,8 +26,8 @@ class PayPalClient_Tests: XCTestCase {
     
     func testVault_whenSandbox_launchesCorrectURLInWebSession() {
         let vaultRequest = PayPalVaultRequest(setupTokenID: "fake-token")
-        payPalClient.vault(vaultRequest)
-        
+        payPalClient.vault(vaultRequest) { _, _ in }
+
         XCTAssertEqual(mockWebAuthenticationSession.lastLaunchedURL?.absoluteString, "https://sandbox.paypal.com/agreements/approve?approval_session_id=fake-token")
     }
     
@@ -40,8 +40,8 @@ class PayPalClient_Tests: XCTestCase {
         )
         
         let vaultRequest = PayPalVaultRequest(setupTokenID: "fake-token")
-        payPalClient.vault(vaultRequest)
-        
+        payPalClient.vault(vaultRequest) { _, _ in }
+
         XCTAssertEqual(mockWebAuthenticationSession.lastLaunchedURL?.absoluteString, "https://paypal.com/agreements/approve?approval_session_id=fake-token")
     }
     
@@ -53,18 +53,16 @@ class PayPalClient_Tests: XCTestCase {
 
         let expectedTokenIDResult = "fakeTokenID"
         let expectedSessionIDResult = "fakeSessionID"
-        let mockVaultDelegate = MockPayPalVaultDelegate(success: {_, result in
-            XCTAssertEqual(expectedTokenIDResult, result.tokenID)
-            XCTAssertEqual(expectedSessionIDResult, result.approvalSessionID)
-            expectation.fulfill()
-        }, error: {_, _ in
-            XCTFail("Invoked error() callback. Should invoke success().")
-        })
-        payPalClient.vaultDelegate = mockVaultDelegate
-        let vaultRequest = PayPalVaultRequest(setupTokenID: "fakeTokenID")
-        payPalClient.vault(vaultRequest)
 
-        waitForExpectations(timeout: 10)
+        let vaultRequest = PayPalVaultRequest(setupTokenID: "fakeTokenID")
+        payPalClient.vault(vaultRequest) { result, error in
+            XCTAssertEqual(expectedTokenIDResult, result?.tokenID)
+            XCTAssertEqual(expectedSessionIDResult, result?.approvalSessionID)
+            XCTAssertNil(error)
+            expectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 2, handler: nil)
     }
 
     func testVault_whenWebSession_cancelled() {
@@ -76,17 +74,41 @@ class PayPalClient_Tests: XCTestCase {
 
         let expectation = expectation(description: "vault(url:) completed")
 
-        let mockVaultDelegate = MockPayPalVaultDelegate(success: {_, _ in
-            XCTFail("Invoked success callback. Should invoke cancel().")
-        }, error: {_, _ in
-            XCTFail("Invoked error() callback. Should invoke success().")
-        }, cancel: { _ in
-            XCTAssert(true)
-            expectation.fulfill()
-        })
-        payPalClient.vaultDelegate = mockVaultDelegate
         let vaultRequest = PayPalVaultRequest(setupTokenID: "fakeTokenID")
-        payPalClient.vault(vaultRequest)
+        payPalClient.vault(vaultRequest) { result, error in
+            XCTAssertNil(result)
+            if let error {
+                XCTAssertEqual(error.domain, PayPalError.domain)
+                XCTAssertEqual(error.code, PayPalError.Code.vaultCanceledError.rawValue)
+                XCTAssertEqual(error.localizedDescription, "PayPal vault has been canceled by the user")
+            } else {
+                XCTFail("Expected error not to be nil")
+            }
+            expectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 10)
+    }
+
+    func testVault_whenWebSession_cancelled_returnsIsVaultCanceledTrue() {
+
+        mockWebAuthenticationSession.cannedErrorResponse = ASWebAuthenticationSessionError(
+            .canceledLogin,
+            userInfo: ["Description": "Mock cancellation error description."]
+        )
+
+        let expectation = expectation(description: "vault(url:) completed")
+
+        let vaultRequest = PayPalVaultRequest(setupTokenID: "fakeTokenID")
+        payPalClient.vault(vaultRequest) { result, error in
+            if let error {
+                XCTAssertNil(result)
+                XCTAssertTrue(PayPalError.isVaultCanceled(error))
+            } else {
+                XCTFail("Expected error from PayPal vault cancellation")
+            }
+            expectation.fulfill()
+        }
 
         waitForExpectations(timeout: 10)
     }
@@ -94,25 +116,25 @@ class PayPalClient_Tests: XCTestCase {
     func testVault_whenWebSession_returnsDefaultError() {
 
         let expectedError = CoreSDKError(
-            code: PayPalWebCheckoutClientError.Code.webSessionError.rawValue,
-            domain: PayPalWebCheckoutClientError.domain,
-            errorDescription: PayPalWebCheckoutClientError.payPalVaultResponseError.errorDescription
+            code: PayPalError.Code.webSessionError.rawValue,
+            domain: PayPalError.domain,
+            errorDescription: PayPalError.payPalVaultResponseError.errorDescription
         )
         mockWebAuthenticationSession.cannedErrorResponse = expectedError
 
         let expectation = expectation(description: "vault(url:) completed")
 
-        let mockVaultDelegate = MockPayPalVaultDelegate(success: {_, _ in
-            XCTFail("Invoked success callback. Should invoke error().")
-        }, error: {_, vaultError in
-            XCTAssertEqual(vaultError.code, expectedError.code)
-            expectation.fulfill()
-        }, cancel: { _ in
-            XCTFail("Invoked cancel callback. Should invoke error().")
-        })
-        payPalClient.vaultDelegate = mockVaultDelegate
         let vaultRequest = PayPalVaultRequest(setupTokenID: "fakeTokenID")
-        payPalClient.vault(vaultRequest)
+        payPalClient.vault(vaultRequest) { result, error in
+            XCTAssertNil(result)
+            if let error {
+                XCTAssertEqual(error.domain, expectedError.domain)
+                XCTAssertEqual(error.code, expectedError.code)
+            } else {
+                XCTFail("Expected error not to be nil")
+            }
+            expectation.fulfill()
+        }
 
         waitForExpectations(timeout: 10)
     }
@@ -124,29 +146,29 @@ class PayPalClient_Tests: XCTestCase {
         let expectation = expectation(description: "vault(url:) completed")
 
         let expectedError = CoreSDKError(
-            code: PayPalWebCheckoutClientError.payPalVaultResponseError.code,
-            domain: PayPalWebCheckoutClientError.domain,
-            errorDescription: PayPalWebCheckoutClientError.payPalVaultResponseError.errorDescription
+            code: PayPalError.payPalVaultResponseError.code,
+            domain: PayPalError.domain,
+            errorDescription: PayPalError.payPalVaultResponseError.errorDescription
         )
 
-        let mockVaultDelegate = MockPayPalVaultDelegate(success: {_, _ in
-            XCTFail("Invoked success() callback. Should invoke error().")
-        }, error: {_, vaultError in
-            XCTAssertEqual(vaultError.code, expectedError.code)
-            expectation.fulfill()
-        })
-        payPalClient.vaultDelegate = mockVaultDelegate
         let vaultRequest = PayPalVaultRequest(setupTokenID: "fakeTokenID")
-        payPalClient.vault(vaultRequest)
+        payPalClient.vault(vaultRequest) { result, error in
+            XCTAssertNil(result)
+            if let error {
+                XCTAssertEqual(error.domain, expectedError.domain)
+                XCTAssertEqual(error.code, expectedError.code)
+            } else {
+                XCTFail("Expected error not to be nil")
+            }
+            expectation.fulfill()
+        }
 
         waitForExpectations(timeout: 10)
     }
 
-    func testStart_whenNativeSDKOnCancelCalled_returnsCancellationError() {
+    func testStart_whenWebAuthenticationSessionCancelCalled_returnsCancellationError() {
         let request = PayPalWebCheckoutRequest(orderID: "1234")
-        let delegate = MockPayPalWebDelegate()
 
-        payPalClient.delegate = delegate
         mockWebAuthenticationSession.cannedErrorResponse = ASWebAuthenticationSessionError(
             _bridgedNSError: NSError(
                 domain: ASWebAuthenticationSessionError.errorDomain,
@@ -155,58 +177,104 @@ class PayPalClient_Tests: XCTestCase {
             )
         )
 
-        payPalClient.start(request: request)
+        let expectation = self.expectation(description: "Call back invoked with error")
+        payPalClient.start(request: request) { result, error in
+            XCTAssertNil(result)
+            if let error {
+                XCTAssertEqual(error.domain, PayPalError.domain)
+                XCTAssertEqual(error.code, PayPalError.checkoutCanceledError.code)
+                XCTAssertEqual(error.localizedDescription, PayPalError.checkoutCanceledError.localizedDescription)
+            } else {
+                XCTFail("Expected error not to be nil")
+            }
+            expectation.fulfill()
+        }
 
-        XCTAssertTrue(delegate.paypalDidCancel)
+        waitForExpectations(timeout: 2, handler: nil)
     }
+
+    func testStart_whenWebSession_cancelled_returnsIsCheckoutCanceledTrue() {
+
+        let request = PayPalWebCheckoutRequest(orderID: "1234")
+
+        mockWebAuthenticationSession.cannedErrorResponse = ASWebAuthenticationSessionError(
+            _bridgedNSError: NSError(
+                domain: ASWebAuthenticationSessionError.errorDomain,
+                code: ASWebAuthenticationSessionError.canceledLogin.rawValue,
+                userInfo: ["Description": "Mock cancellation error description."]
+            )
+        )
+
+        let expectation = self.expectation(description: "Call back invoked with error")
+        payPalClient.start(request: request) { result, error in
+            XCTAssertNil(result)
+            if let error {
+                XCTAssertTrue(PayPalError.isCheckoutCanceled(error))
+            } else {
+                XCTFail("Expected error from PayPal checkout cancellation.")
+            }
+            expectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 2, handler: nil)
+    }
+
 
     func testStart_whenWebAuthenticationSessions_returnsWebSessionError() {
         let request = PayPalWebCheckoutRequest(orderID: "1234")
-        let delegate = MockPayPalWebDelegate()
 
-        payPalClient.delegate = delegate
         mockWebAuthenticationSession.cannedErrorResponse = CoreSDKError(
-            code: PayPalWebCheckoutClientError.Code.webSessionError.rawValue,
-            domain: PayPalWebCheckoutClientError.domain,
+            code: PayPalError.Code.webSessionError.rawValue,
+            domain: PayPalError.domain,
             errorDescription: "Mock web session error description."
         )
 
-        payPalClient.start(request: request)
-
-        let error = delegate.capturedError
-
-        XCTAssertEqual(error?.domain, PayPalWebCheckoutClientError.domain)
-        XCTAssertEqual(error?.code, PayPalWebCheckoutClientError.Code.webSessionError.rawValue)
-        XCTAssertEqual(error?.localizedDescription, "Mock web session error description.")
+        let expectation = self.expectation(description: "Call back invoked with error")
+        payPalClient.start(request: request) { result, error in
+            XCTAssertNil(result)
+            if let error {
+                XCTAssertEqual(error.domain, PayPalError.domain)
+                XCTAssertEqual(error.code, PayPalError.Code.webSessionError.rawValue)
+                XCTAssertEqual(error.localizedDescription, "Mock web session error description.")
+            } else {
+                XCTFail("Expected error not to be nil")
+            }
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: 2, handler: nil)
     }
 
     func testStart_whenResultURLMissingParameters_returnsMalformedResultError() {
         let request = PayPalWebCheckoutRequest(orderID: "1234")
-        let delegate = MockPayPalWebDelegate()
 
-        payPalClient.delegate = delegate
         mockWebAuthenticationSession.cannedResponseURL = URL(string: "https://fakeURL?PayerID=98765")
-        payPalClient.start(request: request)
-
-        let error = delegate.capturedError
-
-        XCTAssertEqual(error?.domain, PayPalWebCheckoutClientError.domain)
-        XCTAssertEqual(error?.code, PayPalWebCheckoutClientError.Code.malformedResultError.rawValue)
-        XCTAssertEqual(error?.localizedDescription, "Result did not contain the expected data.")
+        let expectation = self.expectation(description: "Call back invoked with error")
+        payPalClient.start(request: request) { result, error in
+            XCTAssertNil(result)
+            if let error {
+                XCTAssertEqual(error.domain, PayPalError.domain)
+                XCTAssertEqual(error.code, PayPalError.Code.malformedResultError.rawValue)
+                XCTAssertEqual(error.localizedDescription, "Result did not contain the expected data.")
+            } else {
+                XCTFail("Expected error not to be nil")
+            }
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: 2, handler: nil)
     }
 
     func testStart_whenWebResultIsSuccessful_returnsSuccessfulResult() {
         let request = PayPalWebCheckoutRequest(orderID: "1234")
-        let delegate = MockPayPalWebDelegate()
 
-        payPalClient.delegate = delegate
         mockWebAuthenticationSession.cannedResponseURL = URL(string: "https://fakeURL?token=1234&PayerID=98765")
-        payPalClient.start(request: request)
-
-        let result = delegate.capturedResult
-
-        XCTAssertEqual(result?.orderID, "1234")
-        XCTAssertEqual(result?.payerID, "98765")
+        let expectation = self.expectation(description: "Call back invoked with error")
+        payPalClient.start(request: request) { result, error in
+            XCTAssertEqual(result?.orderID, "1234")
+            XCTAssertEqual(result?.payerID, "98765")
+            XCTAssertNil(error)
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: 2, handler: nil)
     }
 
     func testpayPalCheckoutReturnURL_returnsCorrectURL() {