+++ CHANGELOG.md
@@ -1,6 +1,39 @@
 
 # PayPal iOS SDK Release Notes
 
+## unreleased
+* Breaking Changes
+  * PayPalNativePayments
+    * Remove entire PayPalNativePayments module
+  * PayPalWebPayments
+    * Replace delegate pattern with completion handlers and Swift concurrency
+      * Remove `PayPalWebCheckoutDelegate` and `PayPalVaultDelegate`
+      * Remove `start(request:)` method that uses delegate callbacks
+      * Remove `vault(vaultRequest:)` method that uses delegate callbacks
+      * Add `start(request:completion(PayPalWebCheckoutResult?, CoreSDKError?) -> Void)` to `PayPalWebCheckoutClient`
+      * Add `vault(vaultRequest:completion(PayPalVaultResult?, CoreSDKError?) -> Void)` to `PayPalWebCheckoutClient`
+      * Add `start(request:) async throws -> PayPalCheckoutResult`
+      * Add `vault(vaultRequest:) async throws -> PayPalVaultResult`
+  * CardPayments
+    * Replace delegate pattern with completion handlers and Swift concurrency
+      * Remove `CardDelegate` and `CardVaultDelegate`
+      * Remove `approveOrder(request:)` method that uses delegate callbacks
+      * Remove `vault(vaultRequest:)` method that uses delegate callbacks
+      * Add `approveOrder(request:completion:(CardResult?, CoreSDKError?) -> Void)` to `CardClient`
+      * Add `vault(request:completion:(CardVaultResult?, CoreSDKError?) -> Void)` to `CardClient`
+      * Add `approveOrder(request:) async throws -> CardResult`
+      * Add `vault(vaultRequest:) async throws -> CardVaultResult`   
+* PayPalWebPayments
+  * Rename `PayPalWebCheckoutClientError` to `PayPalError`
+  * Add `.checkoutCanceledError` and `vaultCanceledError` to `PayPalError`
+  * Add public static functions `isCheckoutCanceled(Error)` and `isVaultCanceled(Error)` to `PayPalError` to distinguish cancellation errors in PayPal flows. 
+  * Make `PayPalError` public to expose cancellation error handling helpers
+* CardPayments
+  * Rename `CardClientError` to `CardError`
+  * Add `threeDSecureCanceledError` to `CardError`
+  * Add public static function `isThreeDSecureCanceled(Error)` to `CardError` to distinguish cancellation error from threeDSecure verification
+  * Make `CardError` public to expose cancellation error handling helper
+
 ## 1.5.0 (2024-10-24)
 * PayPalWebPayments
   * Deprecate `PayPalVaultRequest(url:setupTokenID:)`